revert

+++ cosipy/image_deconvolution/dataIF_COSI_DC2.py
@@ -372,15 +372,14 @@ class DataIF_COSI_DC2(ImageDeconvolutionDataInterfaceBase):
         # range to prange()
         # fast math mode turned on
 
-        # loglikelood = np.sum( self.event * np.log(expectation) ) - np.sum(expectation)
-
+        # loglikelood = np.sum( self.event * np.log(expectation) ) - np.sum(expectation)~
         return jit_calc_loglikelihood(self.event.contents.ravel(), expectation.contents.ravel())
 
 @njit(parallel=True, nogil=True, fastmath=True)
 def jit_calc_loglikelihood(event, expectation):
     loglikelihood = 0
     for i in prange(event.shape[0]):
-        loglikelihood += np.log(expectation[i], dtype=np.float32)*event[i] - expectation[i] # np does checks, float64
+        loglikelihood += np.log(expectation[i])*event[i] - expectation[i] # np does checks, float64
         # math? not cmath
         # profile dtype=float64 vs. dtype=np.float32 vs. cmath vs. math
         # different cores
